New proptest reveals Join-Embed-Rename tape bug
(1) I resurrected another part of the old Property Test suite, in which we could express equations like `X|Y == Y|X` and assert them to be true of random grammars X and Y.

Last time, this was expressed by passing in ad-hoc functions for the left and right sides, like the "left" param might be:

    `(x: Grammar, y: Grammar) => Join(x,y)`

This time, you pass in what symbols you want to exist (e.g. `["X", "Y"]`) and then just pass in a Grammar that embeds them:

    `Join(Embed("X"), Embed("Y"))`.

This lets us utilize some of the existing machinery of Collections rather than write it from scratch.

(2) The above test revealed a bug in how we detect and recover from rename errors.  The configuration required to show it is strange and I wouldn't have come up with it myself, but the bug is pretty general.

The issue seems to be, in the second pass of tapecalc, we can catch erroneous renames that we couldn't originally... but by the time we catch them and create an `.ERR` tape to fix things, we've already set in stone other tape decisions that depend on this.

We need to be catching this error earlier.  We couldn't before, actually, because we didn't have error-messaging capability during tape unification.  The previous commit makes that possible, however.

+++ packages/parser/tests/property/reduceGrammar.ts
@@ -3,17 +3,32 @@ import { CollectionGrammar, EpsilonGrammar, Grammar, RenameGrammar, SequenceGram
 import { update } from "../../src/utils/func";
 import { randomChoice, range } from "./randomGrammar";
 
-export function reduceGrammar(g: Grammar): Grammar {
+export type ReduceOptions = {
+    "symbolDrop": boolean
+}
+
+const DEFAULT_OPTIONS: ReduceOptions = {
+    "symbolDrop": true
+}
+
+export function ReduceOptions(
+    opt: Partial<ReduceOptions> = {}
+): ReduceOptions {
+    return { ...DEFAULT_OPTIONS, ...opt };
+}
+
+
+export function reduceGrammar(g: Grammar, opt: ReduceOptions): Grammar {
     switch (g.tag) {
-        case "collection": return reduceCollection(g);
-        case "seq":        return reduceSeq(g);
-        case "rename":     return reduceRename(g);
+        case "collection": return reduceCollection(g, opt);
+        case "seq":        return reduceSeq(g, opt);
+        case "rename":     return reduceRename(g, opt);
     }
     return Epsilon();
 }
 
-export function reduceCollection(g: CollectionGrammar): CollectionGrammar {
-    if (Math.random() < 0.5) {
+export function reduceCollection(g: CollectionGrammar, opt: ReduceOptions): CollectionGrammar {
+    if (opt.symbolDrop && Math.random() < 0.5) {
         // drop one symbol at random
         return dropCollection(g);
     }
@@ -25,7 +40,7 @@ export function reduceCollection(g: CollectionGrammar): CollectionGrammar {
     // simplify one child at random
     const keys = Object.keys(g.symbols);
     const randomKey = randomChoice(keys);
-    result.symbols[randomKey] = reduceGrammar(result.symbols[randomKey]);
+    result.symbols[randomKey] = reduceGrammar(result.symbols[randomKey], opt);
     return result;
 }
 
@@ -39,7 +54,7 @@ export function dropCollection(g: CollectionGrammar): CollectionGrammar {
     return result;
 }
 
-export function reduceSeq(g: SequenceGrammar): Grammar {
+export function reduceSeq(g: SequenceGrammar, opt: ReduceOptions): Grammar {
     if (Math.random() < 0.5) {
         // drop one child at random
         return dropSeq(g);
@@ -53,7 +68,7 @@ export function reduceSeq(g: SequenceGrammar): Grammar {
             newChildren.push(g.children[i]);
             continue;
         }
-        const newChild = reduceGrammar(g.children[i]);
+        const newChild = reduceGrammar(g.children[i], opt);
         newChildren.push(newChild);
         
     }
@@ -87,13 +102,13 @@ function simplifySeq(g: SequenceGrammar): Grammar {
 
 }
 
-export function reduceRename(g: RenameGrammar): Grammar {
+export function reduceRename(g: RenameGrammar, opt: ReduceOptions): Grammar {
     if (Math.random() < 0.5) {
         // drop the renaming
         return g.child;
     }
 
     // otherwise reduce the child
-    const newChild = reduceGrammar(g.child);
+    const newChild = reduceGrammar(g.child, opt);
     return update(g, { child: newChild });
 }
